kloudclt: move it into a seperate package too
