+++ go/build.sh
@@ -20,7 +20,7 @@ services=(
   koding/kites/kontrol
   koding/kites/klient
   koding/kites/kloud
-  koding/kites/kloud/kloudctl
+  github.com/koding/kloudctl
   koding/virt/vmproxy
   koding/virt/vmtool
   koding/overview
