+++ go/src/github.com/koding/kloudctl/command/build.go
@@ -4,9 +4,8 @@ import (
 	"fmt"
 	"time"
 
-	"koding/kites/kloud/kloud"
-
 	"github.com/koding/kite"
+	"github.com/koding/kloud"
 	"github.com/mitchellh/cli"
 )
