+++ go/src/github.com/koding/kloudctl/command/destroy.go
@@ -2,9 +2,9 @@ package command
 
 import (
 	"fmt"
-	"koding/kites/kloud/kloud"
 
 	"github.com/koding/kite"
+	"github.com/koding/kloud"
 	"github.com/mitchellh/cli"
 )
