+++ go/src/github.com/koding/kloudctl/command/ping.go
